requirements.txt
# basic
black
pylint
pytest
kaggle

# jupyter
numpy
panda
matplotlib
seaborn

# sql
sqlalchemy
pandasql

# cli
typer
